+++ src/main/java/unimelb/bitbox/util/RequestState.java
@@ -3,13 +3,13 @@ package unimelb.bitbox.util;
 public class RequestState {
     private String command = null;
     private String pathName = null;
-    private int position = 0;
-    private int length = 0;
+    private long position = 0;
+    private long length = 0;
     public RequestState(String cmd, String pathName){
         this.command = cmd;
         this.pathName = pathName;
     }
-    public RequestState(String cmd, String pathName, int pos, int len){
+    public RequestState(String cmd, String pathName, long pos, long len){
         this.command = cmd;
         this.pathName = pathName;
         this.position = pos;
@@ -21,10 +21,10 @@ public class RequestState {
     public String getPathName(){
         return this.pathName;
     }
-    public int getPosition(){
+    public long getPosition(){
         return this.position;
     }
-    public int getLength(){
+    public long getLength(){
         return this.length;
     }
     public boolean equals(Object obj){
